pyproject.toml
[tool.poetry]
name = "for-students-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
pillow = "^10.4.0"
djangorestframework = "^3.15.2"
requests = "^2.32.3"
aiogram = "^3.13.1"
python-dotenv = "^1.0.1"
uvicorn = "^0.30.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
